refactor(dx): rename orchestrator.py module and classes to dispatcher (#1137)
Rename OrchestratorBridge -> DispatcherBridge, OrchestratorInstruction ->
DispatcherInstruction, create_orchestrator_bridge -> create_dispatcher_bridge,
and build_orchestrator_status -> build_dispatcher_status. Create new
dispatcher.py as the canonical module and convert orchestrator.py to a
backward-compat shim. Also rename scripts/orchestrator-request.py and
scripts/orchestrator-send.py to their dispatcher equivalents.
