+++ packages/telegram-bridge/src/telegram_bridge/__init__.py
@@ -1,31 +1,37 @@
-"""Judgemind Telegram bridge — Python client for agent notifications and commands."""
+"""Judgemind Telegram bridge -- Python client for agent notifications and commands."""
 
 from .client import TelegramBridge
+from .dispatcher import (
+    Command,
+    CommandKind,
+    DispatcherBridge,
+    DispatcherInstruction,
+    InstructionKind,
+    OrchestratorBridge,
+    OrchestratorInstruction,
+    PendingReply,
+    create_dispatcher_bridge,
+    create_orchestrator_bridge,
+)
 from .formatting import escape_html, linkify_github_refs, split_message
 from .interpreter import (
     InterpretedMessage,
     RateLimiter,
     RateLimitError,
+    build_dispatcher_status,
     build_orchestrator_status,
     clear_client_cache,
     get_client,
     interpret_message,
 )
 from .models import Message, SentMessage
-from .orchestrator import (
-    Command,
-    CommandKind,
-    InstructionKind,
-    OrchestratorBridge,
-    OrchestratorInstruction,
-    PendingReply,
-    create_orchestrator_bridge,
-)
 from .validation import ValidationResult, validate_github_links, validate_telegram_payload
 
 __all__ = [
     "Command",
     "CommandKind",
+    "DispatcherBridge",
+    "DispatcherInstruction",
     "InstructionKind",
     "InterpretedMessage",
     "Message",
@@ -37,8 +43,10 @@ __all__ = [
     "SentMessage",
     "TelegramBridge",
     "ValidationResult",
+    "build_dispatcher_status",
     "build_orchestrator_status",
     "clear_client_cache",
+    "create_dispatcher_bridge",
     "create_orchestrator_bridge",
     "get_client",
     "escape_html",
